move more nested components
