--- client/src/components/Menu/index.js
@@ -1,1 +1,0 @@
-export { default } from "./Menu";
